# apps/api/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/account": {
			"post": {
				"description": "Registra uma nova conta bancária para o usuário autenticado. Cada usuário pode ter apenas UMA conta, que é criada automaticamente durante o registro.",
				"operationId": "AccountController_create",
				"parameters": [],
				"requestBody": {
					"required": true,
					"description": "Dados da conta a ser criada",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAccountDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Conta criada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Dados inválidos"
					},
					"401": {
						"description": "Token JWT ausente ou inválido"
					},
					"409": {
						"description": "Usuário já possui uma conta"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Cria uma nova conta bancária",
				"tags": ["account"]
			},
			"get": {
				"description": "Retorna uma lista de todas as contas bancárias do sistema",
				"operationId": "AccountController_findAll",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Lista de contas retornada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/AccountResponseDto"
									}
								}
							}
						}
					},
					"401": {
						"description": "Token JWT ausente ou inválido"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Lista todas as contas",
				"tags": ["account"]
			}
		},
		"/account/me": {
			"get": {
				"description": "Retorna todas as contas bancárias do usuário autenticado via token JWT",
				"operationId": "AccountController_findAllByUser",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Contas do usuário retornadas com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/AccountResponseDto"
									}
								}
							}
						}
					},
					"401": {
						"description": "Token JWT ausente ou inválido"
					},
					"404": {
						"description": "Nenhuma conta encontrada para o usuário"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Lista contas do usuário autenticado",
				"tags": ["account"]
			}
		},
		"/account/me/single": {
			"get": {
				"description": "Retorna a única conta bancária do usuário autenticado. Cada usuário tem apenas uma conta.",
				"operationId": "AccountController_findMe",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Conta do usuário retornada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountResponseDto"
								}
							}
						}
					},
					"401": {
						"description": "Token JWT ausente ou inválido"
					},
					"404": {
						"description": "Conta não encontrada para o usuário"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Retorna a conta única do usuário",
				"tags": ["account"]
			}
		},
		"/account/{id}": {
			"get": {
				"description": "Retorna os dados de uma conta bancária pelo seu UUID",
				"operationId": "AccountController_findOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "UUID da conta a ser recuperada",
						"schema": {
							"example": "550e8400-e29b-41d4-a716-446655440000",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Conta encontrada e retornada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountResponseDto"
								}
							}
						}
					},
					"401": {
						"description": "Token JWT ausente ou inválido"
					},
					"404": {
						"description": "Conta não encontrada com esse UUID"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Obtém uma conta específica",
				"tags": ["account"]
			},
			"patch": {
				"description": "Atualiza os dados de uma conta bancária existente",
				"operationId": "AccountController_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "UUID da conta a ser atualizada",
						"schema": {
							"example": "550e8400-e29b-41d4-a716-446655440000",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"description": "Dados da conta a serem atualizados",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAccountDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Conta atualizada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Dados inválidos"
					},
					"401": {
						"description": "Token JWT ausente ou inválido"
					},
					"404": {
						"description": "Conta não encontrada"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Atualiza uma conta bancária",
				"tags": ["account"]
			},
			"delete": {
				"description": "Remove uma conta bancária do sistema",
				"operationId": "AccountController_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "UUID da conta a ser deletada",
						"schema": {
							"example": "550e8400-e29b-41d4-a716-446655440000",
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Conta deletada com sucesso"
					},
					"401": {
						"description": "Token JWT ausente ou inválido"
					},
					"404": {
						"description": "Conta não encontrada"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Deleta uma conta bancária",
				"tags": ["account"]
			}
		},
		"/account/{id}/loan": {
			"post": {
				"description": "Aprova um empréstimo e atualiza o saldo da conta",
				"operationId": "AccountController_loanRequest",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "UUID da conta para a qual o empréstimo está sendo solicitado",
						"schema": {
							"example": "550e8400-e29b-41d4-a716-446655440000",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Empréstimo aprovado e saldo atualizado com sucesso"
					},
					"401": {
						"description": "Token JWT ausente ou inválido"
					},
					"404": {
						"description": "Conta não encontrada"
					}
				},
				"security": [
					{
						"JWT-auth": []
					}
				],
				"summary": "Solicita um empréstimo para uma conta",
				"tags": ["account"]
			}
		},
...
		"/account-key": {
			"post": {
				"description": "Cria uma nova chave PIX vinculada a uma conta",
				"operationId": "AccountKeyController_create",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAccountKeyDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Chave PIX criada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountKeyResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Chave PIX já cadastrada"
					},
					"404": {
						"description": "Conta não encontrada"
					}
				},
				"summary": "Criar nova chave PIX",
				"tags": ["Account Keys"]
			},
			"get": {
				"description": "Lista todas as chaves PIX vinculadas às contas do usuário",
				"operationId": "AccountKeyController_findAll",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Lista de chaves PIX",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/AccountKeyResponseDto"
									}
								}
							}
						}
					}
				},
				"summary": "Listar chaves PIX do usuário",
				"tags": ["Account Keys"]
			}
		},
		"/account-key/search": {
			"get": {
				"description": "Busca uma chave PIX pelo seu valor e retorna informações da conta vinculada",
				"operationId": "AccountKeyController_findByKey",
				"parameters": [
					{
						"name": "key",
						"required": true,
						"in": "query",
						"description": "Valor da chave PIX (email, CPF, telefone, etc)",
						"schema": {
							"example": "user@example.com",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Chave PIX encontrada com dados da conta",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "string",
											"format": "uuid"
										},
										"key": {
											"type": "string"
										},
										"accountId": {
											"type": "string",
											"format": "uuid"
										},
										"accountNumber": {
											"type": "string"
										},
										"recipientName": {
											"type": "string"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Chave PIX não encontrada"
					}
				},
				"summary": "Buscar chave PIX por valor",
				"tags": ["Account Keys"]
			}
		},
		"/account-key/{id}": {
			"get": {
				"description": "Retorna uma chave PIX específica pelo seu ID",
				"operationId": "AccountKeyController_findOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "UUID da chave PIX",
						"schema": {
							"format": "uuid",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Chave PIX encontrada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountKeyResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Chave PIX não encontrada"
					}
				},
				"summary": "Buscar chave PIX por ID",
				"tags": ["Account Keys"]
			},
			"patch": {
				"description": "Atualiza o valor de uma chave PIX existente",
				"operationId": "AccountKeyController_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "UUID da chave PIX",
						"schema": {
							"format": "uuid",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAccountKeyDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Chave PIX atualizada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountKeyResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Chave PIX não encontrada"
					}
				},
				"summary": "Atualizar chave PIX",
				"tags": ["Account Keys"]
			},
			"delete": {
				"description": "Remove uma chave PIX (exclusão lógica)",
				"operationId": "AccountKeyController_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "UUID da chave PIX",
						"schema": {
							"format": "uuid",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Chave PIX removida com sucesso"
					},
					"404": {
						"description": "Chave PIX não encontrada"
					}
				},
				"summary": "Remover chave PIX",
				"tags": ["Account Keys"]
			}
		}
	},
	"info": {
		"title": "Bank API",
		"description": "API RESTful completa para gerenciamento bancário com suporte a autenticação JWT, operações de contas, transações e gerenciamento de usuários.",
		"version": "1.0.0",
		"contact": {
			"name": "API Support",
			"url": "https://github.com",
			"email": "support@bank-api.com"
		},
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
...
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Desenvolvimento Local"
		},
		{
			"url": "https://api.bank-project.com",
			"description": "Produção"
		}
	],
	"components": {
		"securitySchemes": {
			"JWT-auth": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
...
			"CreateAccountDto": {
				"type": "object",
				"properties": {
					"balance": {
						"type": "number",
						"example": 100000,
						"description": "Saldo inicial da conta em centavos",
						"minimum": 0
					}
				},
				"required": ["balance"]
			},
			"AccountKeyDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "550e8400-e29b-41d4-a716-446655440003",
						"description": "UUID único da chave",
						"format": "uuid"
					},
					"key": {
						"type": "string",
						"example": "user@email.com",
						"description": "Chave PIX"
					},
					"createdAt": {
						"type": "string",
						"example": "2024-10-16T00:00:00Z",
						"description": "Data de criação da chave",
						"format": "date-time"
					}
				},
				"required": ["id", "key", "createdAt"]
			},
			"AccountResponseDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "550e8400-e29b-41d4-a716-446655440000",
						"description": "UUID único da conta",
						"format": "uuid"
					},
					"userId": {
						"type": "string",
						"example": "550e8400-e29b-41d4-a716-446655440001",
						"description": "UUID do usuário proprietário da conta",
						"format": "uuid"
					},
					"accountNumber": {
						"type": "string",
						"example": "123456",
						"description": "Número da conta bancária"
					},
					"balance": {
						"type": "number",
						"example": 100000,
						"description": "Saldo da conta em centavos",
						"minimum": 0
					},
					"accountKeys": {
						"description": "Chaves PIX vinculadas à conta",
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/AccountKeyDto"
						}
					},
					"createdAt": {
						"type": "string",
						"example": "2024-10-16T00:00:00Z",
						"description": "Data de criação da conta",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"example": "2024-10-16T00:00:00Z",
						"description": "Data da última atualização da conta",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"accountNumber",
					"balance",
					"createdAt",
					"updatedAt"
				]
			},
			"UpdateAccountDto": {
				"type": "object",
				"properties": {}
			},
...
			"CreateAccountKeyDto": {
				"type": "object",
				"properties": {
					"accountId": {
						"type": "string",
						"example": "550e8400-e29b-41d4-a716-446655440000",
						"description": "UUID da conta à qual a chave PIX será vinculada",
						"format": "uuid"
					},
					"key": {
						"type": "string",
						"example": "user@email.com",
						"description": "Chave PIX (email, telefone, CPF, CNPJ ou chave aleatória)"
					}
				},
				"required": ["accountId", "key"]
			},
			"AccountKeyResponseDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "550e8400-e29b-41d4-a716-446655440003",
						"description": "UUID único da chave PIX",
						"format": "uuid"
					},
					"key": {
						"type": "string",
						"example": "user@email.com",
						"description": "Chave PIX (email, telefone, CPF, CNPJ ou chave aleatória)"
					},
					"accountId": {
						"type": "string",
						"example": "550e8400-e29b-41d4-a716-446655440000",
						"description": "UUID da conta à qual a chave está vinculada",
						"format": "uuid"
					},
					"createdAt": {
						"type": "string",
						"example": "2024-10-16T00:00:00Z",
						"description": "Data de criação da chave PIX",
						"format": "date-time"
					}
				},
				"required": ["id", "key", "accountId", "createdAt"]
			},
			"UpdateAccountKeyDto": {
				"type": "object",
				"properties": {
					"key": {
						"type": "string",
						"example": "newuser@email.com",
						"description": "Nova chave PIX"
					}
				},
				"required": ["key"]
			}
		}
	},
...
}